update: modificación de modelos, interfaces y repositorios. Se cambio el usuario propio por el de Identity.

+++ GestionTareas_DataAccessLayer/Interfaces/IRepository.cs
@@ -3,7 +3,7 @@
     public interface IRepository<T> where T : class
     {
         Task<IEnumerable<T>> GetAllAsync();
-        Task<T> GetByIdAsync(long id);
+        Task<T> GetByIdAsync(int id);
         Task<bool> AddAsync(T entity);
         Task<bool> UpdateAsync(T entity);
         Task<bool> DeleteAsync(T entity);
