+++ GestionTareas_DataAccessLayer/Repositories/Repository.cs
@@ -18,7 +18,7 @@ namespace GestionTareas_DataAccessLayer.Repositories
             return await _context.Set<T>().ToListAsync();
         }
 
-        public async Task<T> GetByIdAsync(long id)
+        public async Task<T> GetByIdAsync(int id)
         {
             return await _context.Set<T>().FindAsync(id);
         }
